Eliminate definitions filter
We were only showing global definitions within a definitions list, to
avoid showing pop-up definitions within the very definintion from which
that definition was extracted. But not only is the method of
accomplishing this not abstracted adequately, but it's also
inaccurate—in a listing of part definitions, we would want chapter and
title definitions, not just global definitions. Best to remove this
limitation entirely. Per #200.