Eliminate redundant file includes
With hesitation, I'm removing a pair of `require_once()` calls that
@krues8dr added. They're already called within `page-head.inc.php`, so
those calls shouldn't be necessary.

The only scenario in which I can see them being useful is if loading
them is failing via the CLI, because of the non-`dirname()` method of
loading `config.inc.php`. If that's so, then I think we ought to
(obviously) restore those two lines and then provide a comment that
indicates why those lines need to remain there, to avoid future
confusion.