Broaden matching of definition scope indicators
We were missing the definitions in Virginia § 18.2-186.4:1, and it
turned out to be because the scope indicator was "for purposes of this
section," and the list of candidate scope indicators only included "
for purposes of this section"—with a leading space. So I've removed the
leading space from every scope indicator that could potentially be used
at the beginning of a subsection.