Eliminate our elaborate mod_rewrite testing
After all of the fancy testing that we instituted in #273, I've
realized that it now serves no purpose whatsoever. Now that we use
routes, the site depends utterly on the presence of mod_rewrite; it's
not even possible to get to the admin section without it. The fact that
somebody is running the environment test is evidence enough that URL
rewriting is working just fine.

That said, it remains important for folks to be able to informed as to
why the site isn't working. To that end, I have re-opened #351 ("Create
a stand-alone environment test program"). I think we'd just need to
pull out the environment testing class, re-add the URL rewrite testing,
call it from a stand-alone PHP file, and make it a separate project.