Attempt to autoload this class
This evolved from an earlier version, in which @krues8dr was checking
if the file of the class’s existed, so that the class wouldn’t be
instantiated. I replaced it with class_exist(), suppressing its
autoloading functionality. On reflection, that was a mistake—we want it
to try to autoload the class from its file, or so I think, and that
does fix a bug that I was experiencing on Virginia Decoded.

Does this make sense to you, @krues8dr?