Improve the downloads URL parsing
Make two changes. The first is to drop `$` from the end of the
RewriteRule, which mysteriously prevents `/downloads/` from being
passed through `index.php`. But making that looser means that downloads
no longer work. So that’s the reason for the second change, which is
preceding the `RewriteRule` with a `RewriteCond` that keeps it from
taking effect for anything in a subdirectory of `/downloads/` (that is
to say, all downloads).

Closes #551.